refactor: split the reviewer vocabulary out of cosValidation.js (#5702)
`server/lib/cosValidation.js` had two tenants: a Zod schema module and a
code-review domain module. The first ~1150 lines contained no Zod at all —
they were the Review-Loop reviewer roster, alias table, per-reviewer
model/effort ladders, slug→CLI-binary map, pin normalizers/resolvers, and the
builders that render a reviewer set into slashdo's `--review-with` argv and
into the reviewer notes spliced into agent prompts. A contributor looking for
"where does PortOS decide what `agy --model` a reviewer agent runs" had no
reason to open a file named `cosValidation.js`.

That block now lives in `server/lib/reviewerConfig.js` (Zod-free, importing
only the leaf provider/CLI modules it already used). `cosValidation.js` keeps
the CoS schemas and re-exports the new module flat, so all existing
`cosValidation.js` / `validation.js` imports keep resolving unchanged — the
same transitional-shim pattern the #1151/#1831 splits use. The three heaviest
consumers (`promptSections/reviewLifecycle.js`, `promptSections/completion.js`,
`agentWorktreeCleanup.js`) now import from `reviewerConfig.js` directly.

`reviewerConfigMetadata` stayed behind in `cosValidation.js`: it is a thin
wrapper over the Zod-backed `sanitizeTaskMetadata`, and moving it would have
closed an import cycle and dragged Zod back into the reviewer vocabulary.

Tests split alongside the source, plus a new case pinning the transitional
shim — every `reviewerConfig` export must remain reachable as the SAME object
from both `cosValidation.js` and `validation.js`. A dropped re-export fails no
reviewer test (those import the new module directly); it would surface only as
a runtime error inside prompt assembly for a scheduled review-loop task.
