+++ server/lib/slashdoInvocation.js
@@ -58,7 +58,7 @@ export const SLASHDO_INLINE_BUDGET_CHARS = 24000;
  * prompt saving is — pruning to a single CLI reviewer measured -23% on `review`,
  * -27% on `pr`, -28% on `depfree`; the file pointer is the backstop for the rest.
  *
- * Keyed by the PortOS reviewer slug (`REVIEWER_VALUES` in `cosValidation.js`) so
+ * Keyed by the PortOS reviewer slug (`REVIEWER_VALUES` in `reviewerConfig.js`) so
  * the keep-set derives from already-resolved run settings. `copilot` and the
  * arbitrary-`@login` loop are GitHub-side; `claude`/`codex`/`antigravity`/`grok`/`cursor`
  * all share slashdo's one local-agent loop; `ollama`/`lmstudio` are the
@@ -85,10 +85,11 @@ export const SLASHDO_REVIEWER_INCLUDE_NAMES = Object.freeze(Object.values(SLASHD
 /**
  * Reviewer slugs that drive slashdo's shared local-agent (spawnable CLI) loop.
  *
- * Kept here rather than imported from cosValidation.js, which imports THIS
- * module — an import back would be a cycle. Exported so cosValidation.test.js
- * can pin it against `REVIEWER_CLI_BINARIES` (whose keys are the same roster);
- * a reviewer added to one and not the other is a drift the test catches.
+ * Kept here rather than imported from reviewerConfig.js, whose importer
+ * cosValidation.js imports THIS module — an import back would be a cycle.
+ * Exported so reviewerConfig.test.js can pin it against `REVIEWER_CLI_BINARIES`
+ * (whose keys are the same roster); a reviewer added to one and not the other
+ * is a drift the test catches.
  */
 export const LOCAL_AGENT_REVIEWERS = new Set(['claude', 'codex', 'antigravity', 'grok', 'cursor']);
 /** Reviewer slugs that drive slashdo's local-model (Ollama-style) loop. */
